+++ prbt_hardware_support/CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog for package prbt_hardware_support
 
 Forthcoming
 -----------
+* Infinite modbus connection retries
 * Refactor components to use Cartesian speed monitor functionality in controller
 * Contributors: Pilz GmbH and Co. KG